+++ client/client_metadata_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin && !linux
+
+package client
+
+func osVersion() string {
+	return ""
+}